DS-628: Update composer dependencies in bolt_connect to support Twig 2
basaltinc/twig-tools should use the master branch of our fork at
https://github.com/boltdesignsystem/twig-tools.  Updated last in
https://github.com/boltdesignsystem/twig-tools/pull/1

bolt-design-system/core-php should use the latest version of that
project which will be deployed to https://github.com/boltdesignsystem/core-php
(and packagist) when this branch is included in a Bolt release.  It's
annoying circular; we can't actually test that until we cut the release.